// src/BackEnd_Circulo_Lectura/src/interfaces/modelosApp/modelosApp.ts
export interface LibroResumen {
	id_libro: number;
	titulo_libro: string;
	autores: Array<{ nombre_autor: string; apellido_autor: string }>;
	calificacionPromedio?: number;
}
// ...
export interface EventoApp extends EventoBD {
	nombreCreador: string;
}